Handle edge cases around empty results (#291)
In my usage of this plugin, I've come across an issue where cy.react returns an array of empty (null or undefined) value(s). This is happening on a page where the component I'm querying for will not render until a certain threshold of data is loaded into the application. The behavior I expect is that cy.react keeps retrying until the element renders (or until timeout), but in practice, I find that it doesn't retry once, but it immediately returns an array of `undefined`. I suspect that there is an interim state where the component exists in the virtual dom (thus resq can find it), but has not yet been rendered to the dom.

These changes, tested against the application I'm working on, handle this case and allow the plugin to know to retry.